01_etl/requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.1.1
urllib3==1.26.8
